+++ core/version/bump.sh
@@ -11,7 +11,6 @@ ASSETS_DIR="$ROOT_DIR/assets"
 
 # Colors
 GREEN='\033[0;32m'
-YELLOW='\033[1;33m'
 RED='\033[0;31m'
 NC='\033[0m'